Added DjangoRestFramework
To add friends, user should be able to click a button on a Player-page. Added DRF so we can add a little bit of ajax here to patch request.user with added friend.
Needs custom update/partial_update methods to PUT/PATCH friends with just ID, and not full id, url, nickname. Could just use pk in get response ofc, but but it's not good enough!